Refactor game and lineup filters to use a single common component
