+++ src/components/CommonFilter.tsx
@@ -1,0 +1,310 @@
+// React imports:
+import React, { useState, useEffect } from 'react';
+
+// Next imports:
+import { NextPage } from 'next';
+
+// Bootstrap imports:
+import 'bootstrap/dist/css/bootstrap.min.css';
+import Form from 'react-bootstrap/Form';
+import Button from 'react-bootstrap/Button';
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+import Alert from 'react-bootstrap/Alert';
+import InputGroup from 'react-bootstrap/InputGroup';
+
+// Additional components:
+import Select, { components} from "react-select"
+import queryString from "query-string";
+// @ts-ignore
+import LoadingOverlay from 'react-loading-overlay';
+
+// Component imports:
+import { TeamStatsModel } from '../components/TeamStatsTable';
+import { RosterCompareModel } from '../components/RosterCompareTable';
+import { dataLastUpdated } from '../utils/internal-data/dataLastUpdated';
+import { preloadedData } from '../utils/internal-data/preloadedData';
+import { AvailableTeams } from '../utils/internal-data/AvailableTeams';
+import { ClientRequestCache } from '../utils/ClientRequestCache';
+
+// Library imports:
+import fetch from 'isomorphic-unfetch';
+
+export type CommonFilterParams = {
+  year?: string,
+  team?: string,
+  gender?: string,
+  minRank?: string,
+  maxRank?: string
+}
+interface Props<PARAMS> {
+  startingState: PARAMS;
+  onChangeState: (newParams: PARAMS) => void;
+  onChangeCommonState: (newCommonParams: CommonFilterParams) => void;
+  tablePrefix: string,
+  buildParamsFromState: (inHandleResponse: Boolean) => PARAMS;
+  childHandleResponse: (json: any, wasError: Boolean) => void;
+  childSubmitRequest: (paramStr: string, callback: (resp: fetch.IsomorphicResponse) => void) => void;
+}
+
+/** Type workaround per https://stackoverflow.com/questions/51459971/type-of-generic-stateless-component-react-or-extending-generic-function-interfa */
+type CommonFilterI<PARAMS = any> = React.FunctionComponent<Props<PARAMS>>
+
+const CommonFilter: CommonFilterI = ({
+    children,
+    startingState, onChangeState, onChangeCommonState,
+    tablePrefix, buildParamsFromState, childHandleResponse, childSubmitRequest
+}) => {
+
+  // Data model
+
+  // Ugly internals
+  const [ queryIsLoading, setQueryIsLoading ] = useState(false);
+  const [ atLeastOneQueryMade, setAtLeastOneQueryMade ] = useState(false);
+  const [ pageJustLoaded, setPageJustLoaded ] = useState(true);
+  const [ currState, setCurrState ] = useState(startingState);
+
+  // Data source
+  const [ team, setTeam ] = useState(startingState.team || "");
+  const [ year, setYear ] = useState(startingState.year || "2019/20");
+  const [ gender, setGender ] = useState(startingState.gender || "Men");
+  /** Pre-calculate this */
+  const teamList = AvailableTeams.getTeams(null, year, gender);
+
+  // Generic filters:
+
+  const [ minRankFilter, setMinRankFilter ] = useState(startingState.minRank || "0");
+  const [ maxRankFilter, setMaxRankFilter ] = useState(startingState.maxRank || "400");
+
+  // Automatically update child state when any current param is changed:
+  useEffect(() => {
+    onChangeCommonState({
+      team: team, year: year, gender: gender, minRank: minRankFilter, maxRank: maxRankFilter
+    })
+  }, [ team, year, gender, minRankFilter, maxRankFilter ]);
+
+  const [ submitDisabled, setSubmitDisabled ] = useState(false); // (always start as true on page load)
+
+  const isDebug = (process.env.NODE_ENV !== 'production');
+
+  // Utils
+
+  const currentJsonEpoch = dataLastUpdated[`${gender}_${year}`] || -1;
+
+  /** Checks if the input has been changed, and also handles on page load logic */
+  useEffect(() => {
+    setSubmitDisabled(shouldSubmitBeDisabled());
+
+    // Cached reesponse and pre-load handling:
+    if (pageJustLoaded) {
+      setPageJustLoaded(false); //(ensures this code only gets called once)
+
+      const epochRefreshed = ClientRequestCache.refreshEpoch(gender, year, currentJsonEpoch, isDebug);
+
+      // Check for pre-loads:
+      Object.entries(preloadedData || {}).map(function(keyVal) {
+        //TODO: check for prefix?
+        const key = keyVal[0];
+        const valAsJson = keyVal[1];
+        if (!epochRefreshed && ClientRequestCache.peekForResponse(key, "")) {
+          if (isDebug) {
+            console.log(`Already pre-loaded [${key}]`);
+          }
+        } else {
+          if (isDebug) {
+            console.log(`Pre-loading [${key}]`);
+          }
+          ClientRequestCache.cacheResponse(
+            key, "", valAsJson, currentJsonEpoch, isDebug //(no prefix since the key already has it)
+          );
+        }
+      });
+      // Check if object is in cache and onSubmit if so
+      const newParamsStr = queryString.stringify(buildParamsFromState(false));
+      const cachedJson = ClientRequestCache.decacheResponse(
+        newParamsStr, tablePrefix, currentJsonEpoch, isDebug
+      );
+      if (cachedJson) {
+        handleResponse(cachedJson);
+      }
+    }
+  });
+
+  /** If the params match the last request, disable submit */
+  function shouldSubmitBeDisabled() {
+    const newParams = buildParamsFromState(false);
+    const paramsUnchanged = Object.keys(newParams).every(
+      (key: string) => (newParams as any)[key] == (currState as any)[key]
+    );
+    return (atLeastOneQueryMade && paramsUnchanged) || (team == "");
+  }
+
+  /** Whether any of the queries returned an error - we'll treat them all as errors if so */
+  function isResponseError(resp: any) {
+    const jsons = resp?.responses || [];
+    const teamJson = (jsons.length > 0) ? jsons[0] : resp;
+      //(error can be so low level there's not even a responses)
+    const rosterCompareJson = (jsons.length > 1) ? jsons[1] : {};
+    return (Object.keys(teamJson?.error || {}).length > 0) ||
+      (Object.keys(rosterCompareJson?.error || {}).length > 0);
+  }
+
+  /** Handles the response from ES to a stats calc request */
+  function handleResponse(json: any) {
+    setQueryIsLoading(false);
+    const newParams = buildParamsFromState(true); //TODO need to merge optionally
+    const wasError = isResponseError(json);
+    if (!wasError) {
+      setAtLeastOneQueryMade(true);
+      setCurrState(newParams);
+      onChangeState(newParams);
+    }
+    childHandleResponse(json, wasError);
+  }
+
+  /** The user has pressed the submit button - mix of generic and custom logic */
+  function onSubmit() {
+    setQueryIsLoading(true);
+    const newParamsStr = queryString.stringify(buildParamsFromState(false));
+
+    // Check if it's in the cache:
+    const cachedJson = ClientRequestCache.decacheResponse(
+      newParamsStr, tablePrefix, currentJsonEpoch, isDebug
+    );
+    if (cachedJson) {
+      handleResponse(cachedJson);
+    } else {
+      childSubmitRequest(newParamsStr, function(response: fetch.IsomorphicResponse) {
+        response.json().then(function(json: any) {
+          // Cache result locally:
+          if (isDebug) {
+            console.log(`CACHE_KEY=[${tablePrefix}${newParamsStr}]`);
+            console.log(`CACHE_VAL=[${JSON.stringify(json)}]`);
+          }
+          if (!isResponseError(json)) { //(never cache errors)
+            ClientRequestCache.cacheResponse(
+              newParamsStr, tablePrefix, json, currentJsonEpoch, isDebug
+            );
+          }
+          handleResponse(json);
+        })
+      })
+    }
+  }
+
+  /** For use in selects */
+  function stringToOption(s: string) {
+    return { label: s, value: s};
+  }
+  /** For use in team select */
+  function getCurrentTeamOrPlaceholder() {
+    return (team == "") ? { label: 'Choose Team...' } : stringToOption(team);
+  }
+
+  /** Adds the MenuList component with user prompt if there are teams fitered out*/
+  function maybeMenuList() {
+    if (teamList.length < Object.keys(AvailableTeams.byName).length) {
+      return { MenuList };
+    }
+  }
+
+  // Visual components:
+
+  /** Let the user know that he might need to change */
+  const MenuList = (props: any)  => {
+    return (
+      <components.MenuList {...props}>
+        <p className="text-secondary text-center">(Teams filtered by gender/year)</p>
+        {props.children}
+      </components.MenuList>
+    );
+  };
+
+  return <LoadingOverlay
+    active={queryIsLoading}
+    spinner
+    text="Calculating statistics"
+  ><Form>
+    <Form.Group as={Row}>
+      <Col xs={6} sm={6} md={3} lg={2}>
+        <Select
+          value={ stringToOption(gender) }
+          options={Array.from(new Set(AvailableTeams.getTeams(team, year, null).map(
+            (r) => r.gender
+          ))).map(
+            (gender) => stringToOption(gender)
+          )}
+          isSearchable={false}
+          onChange={(option) => { if ((option as any)?.value) setGender((option as any).value) }}
+        />
+      </Col>
+      <Col xs={6} sm={6} md={3} lg={2}>
+        <Select
+          value={ stringToOption(year) }
+          options={Array.from(new Set(AvailableTeams.getTeams(team, null, gender).map(
+            (r) => r.year
+          ))).map(
+            (year) => stringToOption(year)
+          )}
+          isSearchable={false}
+          onChange={(option) => { if ((option as any)?.value) setYear((option as any).value) }}
+        />
+      </Col>
+      <Col className="w-100" bsPrefix="d-lg-none d-md-none"/>
+      <Col xs={12} sm={12} md={6} lg={6}>
+        <Select
+          components = { maybeMenuList() }
+          isClearable={true}
+          styles={{ menu: base => ({ ...base, zIndex: 1000 }) }}
+          value={ getCurrentTeamOrPlaceholder() }
+          options={teamList.map(
+            (r) => stringToOption(r.team)
+          )}
+          onChange={(option) => {
+            setTeam((option as any)?.value || "")
+          }}
+        />
+      </Col>
+    </Form.Group>
+    { children }
+    <Form.Group as={Row} controlId="oppositionFilter">
+      <Form.Label column sm="2">Opponent Strength</Form.Label>
+      <Col sm="2">
+        <InputGroup>
+          <InputGroup.Prepend>
+            <InputGroup.Text id="filterOppoBest">Best</InputGroup.Text>
+          </InputGroup.Prepend>
+          <Form.Control
+            onChange={(ev: any) => {
+              if (ev.target.value.match("^[0-9]*$") != null) {
+                setMinRankFilter(ev.target.value);
+              }
+            }}
+            placeholder = "eg 0"
+            value={minRankFilter}
+          />
+        </InputGroup>
+      </Col>
+        <Col sm="2">
+          <InputGroup>
+            <InputGroup.Prepend>
+              <InputGroup.Text id="filterOppoWorst">Worst</InputGroup.Text>
+            </InputGroup.Prepend>
+            <Form.Control
+              onChange={(ev: any) => {
+                if (ev.target.value.match("^[0-9]*$") != null) {
+                  setMaxRankFilter(ev.target.value);
+                }
+              }}
+              placeholder = "eg 400"
+              value={maxRankFilter}
+            />
+            </InputGroup>
+        </Col>
+      <Form.Label column sm="2">(out of ~360 teams)</Form.Label>
+    </Form.Group>
+    <Button disabled={submitDisabled} variant="primary" onClick={onSubmit}>Submit</Button>
+  </Form></LoadingOverlay>;
+}
+
+export default CommonFilter;
