+++ src/utils/internal-data/preloadedData.ts
@@ -1,23 +1,23 @@
 
 /** A list of key/value parits to preload into local storage - note 'cacheEpoch' is overwritten */
 export const preloadedData: Record<string, any> = {
-  'autoOffQuery=true&baseQuery=&gender=Women&maxRank=400&minRank=0&offQuery=NOT%20%28Kaila%29&onQuery=Kaila&team=Maryland&year=2018%2F9'
+  'game-autoOffQuery=true&baseQuery=&gender=Women&maxRank=400&minRank=0&offQuery=NOT%20%28Kaila%29&onQuery=Kaila&team=Maryland&year=2018%2F9'
   :
   {"cacheEpoch":1572354350,"took":56,"responses":[{"took":56,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":738,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":738,"total_off_fga":{"value":2074},"total_def_2prim_attempts":{"value":618},"total_off_2pmid_attempts":{"value":771},"total_off_2pmid_made":{"value":312},"total_def_fta":{"value":508},"off_2pmidr":{"value":0.3717454194792671},"off_ppp":{"value":105.85548172757476},"off_3pr":{"value":0.2550626808100289},"total_def_fga":{"value":2100},"def_poss":{"value":2420},"total_off_2prim_made":{"value":458},"def_to":{"value":0.21115702479338844},"def_adj_opp":{"value":87.77470930232558},"off_3p":{"value":0.3591682419659735},"def_2p":{"value":0.3815612382234186},"off_2primr":{"value":0.37319189971070393},"def_2primr":{"value":0.29428571428571426},"points_scored":{"value":2549},"off_orb":{"value":0.38218151540383016},"off_adj_opp":{"value":100.21086776859505},"total_def_2pmid_attempts":{"value":868},"def_ppp":{"value":85.99173553719008},"def_2pmidr":{"value":0.41333333333333333},"def_3pr":{"value":0.2923809523809524},"points_allowed":{"value":2080},"def_3p":{"value":0.3241042345276873},"off_poss":{"value":2408},"total_off_2prim_attempts":{"value":774},"def_orb":{"value":0.3012640449438202},"total_def_2pmid_made":{"value":257},"total_def_2prim_made":{"value":310},"off_2p":{"value":0.49838187702265374},"off_efg":{"value":0.508678881388621},"off_to":{"value":0.21553156146179403},"total_off_fta":{"value":628},"def_efg":{"value":0.41214285714285714},"def_2prim":{"value":0.5016181229773463},"off_2pmid":{"value":0.4046692607003891},"def_2pmid":{"value":0.2960829493087558},"off_ftr":{"value":0.30279652844744454},"def_ftr":{"value":0.2419047619047619},"off_2prim":{"value":0.5917312661498708}},"off":{"doc_count":208,"total_off_fga":{"value":523},"total_def_2prim_attempts":{"value":159},"total_off_2pmid_attempts":{"value":175},"total_off_2pmid_made":{"value":66},"total_def_fta":{"value":132},"off_2pmidr":{"value":0.33460803059273425},"off_ppp":{"value":106.54664484451719},"off_3pr":{"value":0.29636711281070743},"total_def_fga":{"value":562},"def_poss":{"value":614},"total_off_2prim_made":{"value":111},"def_to":{"value":0.17263843648208468},"def_adj_opp":{"value":89.27872340425532},"off_3p":{"value":0.4258064516129032},"def_2p":{"value":0.3643410852713178},"off_2primr":{"value":0.3690248565965583},"def_2primr":{"value":0.28291814946619215},"points_scored":{"value":651},"off_orb":{"value":0.38387096774193546},"off_adj_opp":{"value":97.86335504885993},"total_def_2pmid_attempts":{"value":228},"def_ppp":{"value":89.41368078175896},"def_2pmidr":{"value":0.40569395017793597},"def_3pr":{"value":0.3113879003558719},"points_allowed":{"value":549},"def_3p":{"value":0.3485714285714286},"off_poss":{"value":611},"total_off_2prim_attempts":{"value":193},"def_orb":{"value":0.29591836734693877},"total_def_2pmid_made":{"value":68},"total_def_2prim_made":{"value":73},"off_2p":{"value":0.48097826086956524},"off_efg":{"value":0.5277246653919694},"off_to":{"value":0.22913256955810146},"total_off_fta":{"value":156},"def_efg":{"value":0.41370106761565834},"def_2prim":{"value":0.4591194968553459},"off_2pmid":{"value":0.37714285714285717},"def_2pmid":{"value":0.2982456140350877},"off_ftr":{"value":0.2982791586998088},"def_ftr":{"value":0.23487544483985764},"off_2prim":{"value":0.5751295336787565}},"on":{"doc_count":530,"total_off_fga":{"value":1551},"total_def_2prim_attempts":{"value":459},"total_off_2pmid_attempts":{"value":596},"total_off_2pmid_made":{"value":246},"total_def_fta":{"value":376},"off_2pmidr":{"value":0.3842682140554481},"off_ppp":{"value":105.62047857540345},"off_3pr":{"value":0.24113475177304963},"total_def_fga":{"value":1538},"def_poss":{"value":1806},"total_off_2prim_made":{"value":347},"def_to":{"value":0.22425249169435216},"def_adj_opp":{"value":87.26332776850305},"off_3p":{"value":0.3315508021390374},"def_2p":{"value":0.38762511373976344},"off_2primr":{"value":0.37459703417150225},"def_2primr":{"value":0.29843953185955785},"points_scored":{"value":1898},"off_orb":{"value":0.38159371492704824},"off_adj_opp":{"value":101.00897009966778},"total_def_2pmid_attempts":{"value":640},"def_ppp":{"value":84.82834994462901},"def_2pmidr":{"value":0.41612483745123535},"def_3pr":{"value":0.28543563068920674},"points_allowed":{"value":1531},"def_3p":{"value":0.3143507972665148},"off_poss":{"value":1797},"total_off_2prim_attempts":{"value":581},"def_orb":{"value":0.30329457364341084},"total_def_2pmid_made":{"value":189},"total_def_2prim_made":{"value":237},"off_2p":{"value":0.5038232795242141},"off_efg":{"value":0.5022566086395873},"off_to":{"value":0.2109070673344463},"total_off_fta":{"value":472},"def_efg":{"value":0.4115734720416125},"def_2prim":{"value":0.5163398692810458},"off_2pmid":{"value":0.412751677852349},"def_2pmid":{"value":0.2953125},"off_ftr":{"value":0.3043197936814958},"def_ftr":{"value":0.2444733420026008},"off_2prim":{"value":0.5972461273666093}}}}},"status":200},{"took":21,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":738,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":738,"global_poss_count":{"value":2408},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Mikesell, Taylor","doc_count":624,"poss_count":{"value":2115}},{"key":"Charles, Kaila","doc_count":530,"poss_count":{"value":1797}},{"key":"Watson, Blair","doc_count":476,"poss_count":{"value":1679}},{"key":"Jones, Stephanie","doc_count":474,"poss_count":{"value":1608}},{"key":"Lewis, Channise","doc_count":493,"poss_count":{"value":1573}},{"key":"Austin, Shakira","doc_count":433,"poss_count":{"value":1412}},{"key":"Fraser, Brianna","doc_count":295,"poss_count":{"value":902}},{"key":"Vujacic, Sara","doc_count":234,"poss_count":{"value":616}},{"key":"Myers, Sarah","doc_count":91,"poss_count":{"value":228}},{"key":"Owens, Olivia","doc_count":40,"poss_count":{"value":110}}]}},"off":{"doc_count":208,"global_poss_count":{"value":611},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Mikesell, Taylor","doc_count":170,"poss_count":{"value":519}},{"key":"Austin, Shakira","doc_count":155,"poss_count":{"value":466}},{"key":"Lewis, Channise","doc_count":159,"poss_count":{"value":454}},{"key":"Jones, Stephanie","doc_count":144,"poss_count":{"value":433}},{"key":"Watson, Blair","doc_count":126,"poss_count":{"value":401}},{"key":"Vujacic, Sara","doc_count":115,"poss_count":{"value":311}},{"key":"Fraser, Brianna","doc_count":84,"poss_count":{"value":238}},{"key":"Myers, Sarah","doc_count":56,"poss_count":{"value":150}},{"key":"Owens, Olivia","doc_count":31,"poss_count":{"value":83}}]}},"on":{"doc_count":530,"global_poss_count":{"value":1797},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Charles, Kaila","doc_count":530,"poss_count":{"value":1797}},{"key":"Mikesell, Taylor","doc_count":454,"poss_count":{"value":1596}},{"key":"Watson, Blair","doc_count":350,"poss_count":{"value":1278}},{"key":"Jones, Stephanie","doc_count":330,"poss_count":{"value":1175}},{"key":"Lewis, Channise","doc_count":334,"poss_count":{"value":1119}},{"key":"Austin, Shakira","doc_count":278,"poss_count":{"value":946}},{"key":"Fraser, Brianna","doc_count":211,"poss_count":{"value":664}},{"key":"Vujacic, Sara","doc_count":119,"poss_count":{"value":305}},{"key":"Myers, Sarah","doc_count":35,"poss_count":{"value":78}},{"key":"Owens, Olivia","doc_count":9,"poss_count":{"value":27}}]}}}}},"status":200}]}
   ,
-  'autoOffQuery=true&baseQuery=&gender=Men&maxRank=400&minRank=0&offQuery=NOT%20%28Bruno%29&onQuery=Bruno&team=Maryland&year=2018%2F9'
+  'game-autoOffQuery=true&baseQuery=&gender=Men&maxRank=400&minRank=0&offQuery=NOT%20%28Bruno%29&onQuery=Bruno&team=Maryland&year=2018%2F9'
   :
   {"took":114,"responses":[{"took":114,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1424,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1424,"total_off_fga":{"value":3790},"total_def_2prim_attempts":{"value":1032},"total_off_2pmid_attempts":{"value":1146},"total_off_2pmid_made":{"value":472},"total_def_fta":{"value":1046},"off_2pmidr":{"value":0.30237467018469655},"off_ppp":{"value":108.17242927705433},"off_3pr":{"value":0.36939313984168864},"total_def_fga":{"value":3966},"def_poss":{"value":4446},"total_off_2prim_made":{"value":746},"def_to":{"value":0.14260008996851103},"def_adj_opp":{"value":98.67943421643467},"off_3p":{"value":0.3485714285714286},"def_2p":{"value":0.4475873544093178},"off_2primr":{"value":0.32823218997361475},"def_2primr":{"value":0.26021180030257185},"points_scored":{"value":4818},"off_orb":{"value":0.3363148479427549},"off_adj_opp":{"value":110.20620782726046},"total_def_2pmid_attempts":{"value":1372},"def_ppp":{"value":98.740440845704},"def_2pmidr":{"value":0.345940494200706},"def_3pr":{"value":0.39384770549672216},"points_allowed":{"value":4390},"def_3p":{"value":0.3175416133162612},"off_poss":{"value":4454},"total_off_2prim_attempts":{"value":1244},"def_orb":{"value":0.2440944881889764},"total_def_2pmid_made":{"value":506},"total_def_2prim_made":{"value":570},"off_2p":{"value":0.5096234309623431},"off_efg":{"value":0.5145118733509235},"off_to":{"value":0.1948810058374495},"total_off_fta":{"value":1234},"def_efg":{"value":0.4589006555723651},"def_2prim":{"value":0.5523255813953488},"off_2pmid":{"value":0.4118673647469459},"def_2pmid":{"value":0.3688046647230321},"off_ftr":{"value":0.32559366754617414},"def_ftr":{"value":0.26374180534543623},"off_2prim":{"value":0.5996784565916399}},"off":{"doc_count":492,"total_off_fga":{"value":946},"total_def_2prim_attempts":{"value":300},"total_off_2pmid_attempts":{"value":306},"total_off_2pmid_made":{"value":114},"total_def_fta":{"value":316},"off_2pmidr":{"value":0.32346723044397463},"off_ppp":{"value":100},"off_3pr":{"value":0.3657505285412262},"total_def_fga":{"value":1014},"def_poss":{"value":1158},"total_off_2prim_made":{"value":158},"def_to":{"value":0.15544041450777202},"def_adj_opp":{"value":100.14293193717278},"off_3p":{"value":0.3468208092485549},"def_2p":{"value":0.45302013422818793},"off_2primr":{"value":0.3107822410147992},"def_2primr":{"value":0.2958579881656805},"points_scored":{"value":1146},"off_orb":{"value":0.3289036544850498},"off_adj_opp":{"value":108.56373056994819},"total_def_2pmid_attempts":{"value":296},"def_ppp":{"value":102.76338514680484},"def_2pmidr":{"value":0.29191321499013806},"def_3pr":{"value":0.41222879684418146},"points_allowed":{"value":1190},"def_3p":{"value":0.3444976076555024},"off_poss":{"value":1146},"total_off_2prim_attempts":{"value":294},"def_orb":{"value":0.2782874617737003},"total_def_2pmid_made":{"value":94},"total_def_2prim_made":{"value":176},"off_2p":{"value":0.4533333333333333},"off_efg":{"value":0.47780126849894294},"off_to":{"value":0.22164048865619546},"total_off_fta":{"value":328},"def_efg":{"value":0.47928994082840237},"def_2prim":{"value":0.5866666666666667},"off_2pmid":{"value":0.37254901960784315},"def_2pmid":{"value":0.31756756756756754},"off_ftr":{"value":0.346723044397463},"def_ftr":{"value":0.3116370808678501},"off_2prim":{"value":0.5374149659863946}},"on":{"doc_count":932,"total_off_fga":{"value":2844},"total_def_2prim_attempts":{"value":732},"total_off_2pmid_attempts":{"value":840},"total_off_2pmid_made":{"value":358},"total_def_fta":{"value":730},"off_2pmidr":{"value":0.29535864978902954},"off_ppp":{"value":111.00362756952842},"off_3pr":{"value":0.37060478199718705},"total_def_fga":{"value":2952},"def_poss":{"value":3288},"total_off_2prim_made":{"value":588},"def_to":{"value":0.1380778588807786},"def_adj_opp":{"value":98.17243047158405},"off_3p":{"value":0.349146110056926},"def_2p":{"value":0.44579646017699115},"off_2primr":{"value":0.3340365682137834},"def_2primr":{"value":0.24796747967479674},"points_scored":{"value":3672},"off_orb":{"value":0.3390452876376989},"off_adj_opp":{"value":110.78467153284672},"total_def_2pmid_attempts":{"value":1076},"def_ppp":{"value":97.32360097323601},"def_2pmidr":{"value":0.36449864498644985},"def_3pr":{"value":0.3875338753387534},"points_allowed":{"value":3200},"def_3p":{"value":0.3076923076923077},"off_poss":{"value":3308},"total_off_2prim_attempts":{"value":950},"def_orb":{"value":0.23223753976670203},"total_def_2pmid_made":{"value":412},"total_def_2prim_made":{"value":394},"off_2p":{"value":0.5284916201117319},"off_efg":{"value":0.5267229254571026},"off_to":{"value":0.18561064087061668},"total_off_fta":{"value":906},"def_efg":{"value":0.4518970189701897},"def_2prim":{"value":0.5382513661202186},"off_2pmid":{"value":0.4261904761904762},"def_2pmid":{"value":0.3828996282527881},"off_ftr":{"value":0.31856540084388185},"def_ftr":{"value":0.247289972899729},"off_2prim":{"value":0.6189473684210526}}}}},"status":200},{"took":23,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1424,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1424,"global_poss_count":{"value":4454},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":1174,"poss_count":{"value":3888}},{"key":"Fernando, Bruno","doc_count":932,"poss_count":{"value":3308}},{"key":"Ayala, Eric","doc_count":978,"poss_count":{"value":3210}},{"key":"Smith, Jalen","doc_count":822,"poss_count":{"value":2898}},{"key":"Morsell, Darryl","doc_count":814,"poss_count":{"value":2882}},{"key":"Wiggins, Aaron","doc_count":940,"poss_count":{"value":2644}},{"key":"Smith Jr., Serrel","doc_count":550,"poss_count":{"value":1370}},{"key":"Lindo Jr., Ricky","doc_count":558,"poss_count":{"value":1324}},{"key":"Bender, Ivan","doc_count":140,"poss_count":{"value":318}},{"key":"Tomaic, Joshua","doc_count":94,"poss_count":{"value":216}},{"key":"Mona, Reese","doc_count":56,"poss_count":{"value":108}},{"key":"Valmon, Travis","doc_count":20,"poss_count":{"value":32}},{"key":"Terrell, Andrew","doc_count":18,"poss_count":{"value":32}},{"key":"Ramsey, Trace","doc_count":18,"poss_count":{"value":28}},{"key":"Clark, Will","doc_count":6,"poss_count":{"value":12}}]}},"off":{"doc_count":492,"global_poss_count":{"value":1146},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":368,"poss_count":{"value":898}},{"key":"Wiggins, Aaron","doc_count":348,"poss_count":{"value":858}},{"key":"Lindo Jr., Ricky","doc_count":316,"poss_count":{"value":732}},{"key":"Smith, Jalen","doc_count":310,"poss_count":{"value":732}},{"key":"Ayala, Eric","doc_count":294,"poss_count":{"value":682}},{"key":"Smith Jr., Serrel","doc_count":266,"poss_count":{"value":608}},{"key":"Morsell, Darryl","doc_count":232,"poss_count":{"value":544}},{"key":"Bender, Ivan","doc_count":136,"poss_count":{"value":316}},{"key":"Tomaic, Joshua","doc_count":82,"poss_count":{"value":174}},{"key":"Mona, Reese","doc_count":46,"poss_count":{"value":82}},{"key":"Valmon, Travis","doc_count":20,"poss_count":{"value":32}},{"key":"Terrell, Andrew","doc_count":18,"poss_count":{"value":32}},{"key":"Ramsey, Trace","doc_count":18,"poss_count":{"value":28}},{"key":"Clark, Will","doc_count":6,"poss_count":{"value":12}}]}},"on":{"doc_count":932,"global_poss_count":{"value":3308},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Fernando, Bruno","doc_count":932,"poss_count":{"value":3308}},{"key":"Cowan, Anthony","doc_count":806,"poss_count":{"value":2990}},{"key":"Ayala, Eric","doc_count":684,"poss_count":{"value":2528}},{"key":"Morsell, Darryl","doc_count":582,"poss_count":{"value":2338}},{"key":"Smith, Jalen","doc_count":512,"poss_count":{"value":2166}},{"key":"Wiggins, Aaron","doc_count":592,"poss_count":{"value":1786}},{"key":"Smith Jr., Serrel","doc_count":284,"poss_count":{"value":762}},{"key":"Lindo Jr., Ricky","doc_count":242,"poss_count":{"value":592}},{"key":"Tomaic, Joshua","doc_count":12,"poss_count":{"value":42}},{"key":"Mona, Reese","doc_count":10,"poss_count":{"value":26}},{"key":"Bender, Ivan","doc_count":4,"poss_count":{"value":2}}]}}}}},"status":200}]}
   ,
-  'autoOffQuery=true&baseQuery=NOT%20%28Terrell%20OR%20Valmon%20OR%20Clark%29&gender=Men&maxRank=400&minRank=0&offQuery=NOT%20%28Bruno%29&onQuery=Bruno&team=Maryland&year=2018%2F9'
+  'game-autoOffQuery=true&baseQuery=NOT%20%28Terrell%20OR%20Valmon%20OR%20Clark%29&gender=Men&maxRank=400&minRank=0&offQuery=NOT%20%28Bruno%29&onQuery=Bruno&team=Maryland&year=2018%2F9'
   :
   {"took":47,"responses":[{"took":47,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1424,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1400,"total_off_fga":{"value":3768},"total_def_2prim_attempts":{"value":1020},"total_off_2pmid_attempts":{"value":1136},"total_off_2pmid_made":{"value":466},"total_def_fta":{"value":1026},"off_2pmidr":{"value":0.30148619957537154},"off_ppp":{"value":108.55978260869566},"off_3pr":{"value":0.36889596602972397},"total_def_fga":{"value":3936},"def_poss":{"value":4406},"total_off_2prim_made":{"value":746},"def_to":{"value":0.1425329096686337},"def_adj_opp":{"value":98.6243206521739},"off_3p":{"value":0.3496402877697842},"def_2p":{"value":0.4459346186085499},"off_2primr":{"value":0.32961783439490444},"def_2primr":{"value":0.25914634146341464},"points_scored":{"value":4794},"off_orb":{"value":0.33753375337533753},"off_adj_opp":{"value":110.26268724466637},"total_def_2pmid_attempts":{"value":1366},"def_ppp":{"value":98.59282796187017},"def_2pmidr":{"value":0.3470528455284553},"def_3pr":{"value":0.3938008130081301},"points_allowed":{"value":4344},"def_3p":{"value":0.3174193548387097},"off_poss":{"value":4416},"total_off_2prim_attempts":{"value":1242},"def_orb":{"value":0.24425059476605868},"total_def_2pmid_made":{"value":504},"total_def_2prim_made":{"value":560},"off_2p":{"value":0.5096719932716569},"off_efg":{"value":0.5151273885350318},"off_to":{"value":0.19384057971014493},"total_off_fta":{"value":1222},"def_efg":{"value":0.4578252032520325},"def_2prim":{"value":0.5490196078431373},"off_2pmid":{"value":0.4102112676056338},"def_2pmid":{"value":0.3689604685212299},"off_ftr":{"value":0.3243099787685775},"def_ftr":{"value":0.2606707317073171},"off_2prim":{"value":0.6006441223832528}},"off":{"doc_count":468,"total_off_fga":{"value":924},"total_def_2prim_attempts":{"value":288},"total_off_2pmid_attempts":{"value":296},"total_off_2pmid_made":{"value":108},"total_def_fta":{"value":296},"off_2pmidr":{"value":0.3203463203463203},"off_ppp":{"value":101.26353790613719},"off_3pr":{"value":0.36363636363636365},"total_def_fga":{"value":984},"def_poss":{"value":1118},"total_off_2prim_made":{"value":158},"def_to":{"value":0.1556350626118068},"def_adj_opp":{"value":99.97346570397113},"off_3p":{"value":0.35119047619047616},"def_2p":{"value":0.4463667820069204},"off_2primr":{"value":0.31601731601731603},"def_2primr":{"value":0.2926829268292683},"points_scored":{"value":1122},"off_orb":{"value":0.3333333333333333},"off_adj_opp":{"value":108.72754919499107},"total_def_2pmid_attempts":{"value":290},"def_ppp":{"value":102.32558139534883},"def_2pmidr":{"value":0.29471544715447157},"def_3pr":{"value":0.41260162601626016},"points_allowed":{"value":1144},"def_3p":{"value":0.3448275862068966},"off_poss":{"value":1108},"total_off_2prim_attempts":{"value":292},"def_orb":{"value":0.279874213836478},"total_def_2pmid_made":{"value":92},"total_def_2prim_made":{"value":166},"off_2p":{"value":0.4523809523809524},"off_efg":{"value":0.47943722943722944},"off_to":{"value":0.2184115523465704},"total_off_fta":{"value":316},"def_efg":{"value":0.47560975609756095},"def_2prim":{"value":0.5763888888888888},"off_2pmid":{"value":0.36486486486486486},"def_2pmid":{"value":0.31724137931034485},"off_ftr":{"value":0.341991341991342},"def_ftr":{"value":0.3008130081300813},"off_2prim":{"value":0.541095890410959}},"on":{"doc_count":932,"total_off_fga":{"value":2844},"total_def_2prim_attempts":{"value":732},"total_off_2pmid_attempts":{"value":840},"total_off_2pmid_made":{"value":358},"total_def_fta":{"value":730},"off_2pmidr":{"value":0.29535864978902954},"off_ppp":{"value":111.00362756952842},"off_3pr":{"value":0.37060478199718705},"total_def_fga":{"value":2952},"def_poss":{"value":3288},"total_off_2prim_made":{"value":588},"def_to":{"value":0.1380778588807786},"def_adj_opp":{"value":98.17243047158405},"off_3p":{"value":0.349146110056926},"def_2p":{"value":0.44579646017699115},"off_2primr":{"value":0.3340365682137834},"def_2primr":{"value":0.24796747967479674},"points_scored":{"value":3672},"off_orb":{"value":0.3390452876376989},"off_adj_opp":{"value":110.78467153284672},"total_def_2pmid_attempts":{"value":1076},"def_ppp":{"value":97.32360097323601},"def_2pmidr":{"value":0.36449864498644985},"def_3pr":{"value":0.3875338753387534},"points_allowed":{"value":3200},"def_3p":{"value":0.3076923076923077},"off_poss":{"value":3308},"total_off_2prim_attempts":{"value":950},"def_orb":{"value":0.23223753976670203},"total_def_2pmid_made":{"value":412},"total_def_2prim_made":{"value":394},"off_2p":{"value":0.5284916201117319},"off_efg":{"value":0.5267229254571026},"off_to":{"value":0.18561064087061668},"total_off_fta":{"value":906},"def_efg":{"value":0.4518970189701897},"def_2prim":{"value":0.5382513661202186},"off_2pmid":{"value":0.4261904761904762},"def_2pmid":{"value":0.3828996282527881},"off_ftr":{"value":0.31856540084388185},"def_ftr":{"value":0.247289972899729},"off_2prim":{"value":0.6189473684210526}}}}},"status":200},{"took":30,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1424,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1400,"global_poss_count":{"value":4416},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":1174,"poss_count":{"value":3888}},{"key":"Fernando, Bruno","doc_count":932,"poss_count":{"value":3308}},{"key":"Ayala, Eric","doc_count":974,"poss_count":{"value":3204}},{"key":"Smith, Jalen","doc_count":820,"poss_count":{"value":2894}},{"key":"Morsell, Darryl","doc_count":812,"poss_count":{"value":2878}},{"key":"Wiggins, Aaron","doc_count":940,"poss_count":{"value":2644}},{"key":"Smith Jr., Serrel","doc_count":540,"poss_count":{"value":1360}},{"key":"Lindo Jr., Ricky","doc_count":558,"poss_count":{"value":1324}},{"key":"Bender, Ivan","doc_count":130,"poss_count":{"value":304}},{"key":"Tomaic, Joshua","doc_count":78,"poss_count":{"value":194}},{"key":"Mona, Reese","doc_count":40,"poss_count":{"value":82}},{"key":"Ramsey, Trace","doc_count":2,"poss_count":{"value":0}}]}},"off":{"doc_count":468,"global_poss_count":{"value":1108},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":368,"poss_count":{"value":898}},{"key":"Wiggins, Aaron","doc_count":348,"poss_count":{"value":858}},{"key":"Lindo Jr., Ricky","doc_count":316,"poss_count":{"value":732}},{"key":"Smith, Jalen","doc_count":308,"poss_count":{"value":728}},{"key":"Ayala, Eric","doc_count":290,"poss_count":{"value":676}},{"key":"Smith Jr., Serrel","doc_count":256,"poss_count":{"value":598}},{"key":"Morsell, Darryl","doc_count":230,"poss_count":{"value":540}},{"key":"Bender, Ivan","doc_count":126,"poss_count":{"value":302}},{"key":"Tomaic, Joshua","doc_count":66,"poss_count":{"value":152}},{"key":"Mona, Reese","doc_count":30,"poss_count":{"value":56}},{"key":"Ramsey, Trace","doc_count":2,"poss_count":{"value":0}}]}},"on":{"doc_count":932,"global_poss_count":{"value":3308},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Fernando, Bruno","doc_count":932,"poss_count":{"value":3308}},{"key":"Cowan, Anthony","doc_count":806,"poss_count":{"value":2990}},{"key":"Ayala, Eric","doc_count":684,"poss_count":{"value":2528}},{"key":"Morsell, Darryl","doc_count":582,"poss_count":{"value":2338}},{"key":"Smith, Jalen","doc_count":512,"poss_count":{"value":2166}},{"key":"Wiggins, Aaron","doc_count":592,"poss_count":{"value":1786}},{"key":"Smith Jr., Serrel","doc_count":284,"poss_count":{"value":762}},{"key":"Lindo Jr., Ricky","doc_count":242,"poss_count":{"value":592}},{"key":"Tomaic, Joshua","doc_count":12,"poss_count":{"value":42}},{"key":"Mona, Reese","doc_count":10,"poss_count":{"value":26}},{"key":"Bender, Ivan","doc_count":4,"poss_count":{"value":2}}]}}}}},"status":200}]}
   ,
-  'autoOffQuery=true&baseQuery=NOT%20%28Terrell%20OR%20Valmon%20OR%20Clark%29&gender=Men&maxRank=100&minRank=0&offQuery=NOT%20%28Bruno%29&onQuery=Bruno&team=Maryland&year=2018%2F9'
+  'game-autoOffQuery=true&baseQuery=NOT%20%28Terrell%20OR%20Valmon%20OR%20Clark%29&gender=Men&maxRank=100&minRank=0&offQuery=NOT%20%28Bruno%29&onQuery=Bruno&team=Maryland&year=2018%2F9'
   :
   {"took":63,"responses":[{"took":63,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1068,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1062,"total_off_fga":{"value":2830},"total_def_2prim_attempts":{"value":740},"total_off_2pmid_attempts":{"value":888},"total_off_2pmid_made":{"value":362},"total_def_fta":{"value":842},"off_2pmidr":{"value":0.3137809187279152},"off_ppp":{"value":105.60916767189384},"off_3pr":{"value":0.3710247349823322},"total_def_fga":{"value":2988},"def_poss":{"value":3308},"total_off_2prim_made":{"value":506},"def_to":{"value":0.12454655380894801},"def_adj_opp":{"value":95.65144752714114},"off_3p":{"value":0.3523809523809524},"def_2p":{"value":0.45251396648044695},"off_2primr":{"value":0.3151943462897526},"def_2primr":{"value":0.24765729585006693},"points_scored":{"value":3502},"off_orb":{"value":0.324736225087925},"off_adj_opp":{"value":113.00719467956469},"total_def_2pmid_attempts":{"value":1050},"def_ppp":{"value":101.93470374848852},"def_2pmidr":{"value":0.3514056224899598},"def_3pr":{"value":0.4009370816599732},"points_allowed":{"value":3372},"def_3p":{"value":0.31886477462437396},"off_poss":{"value":3316},"total_off_2prim_attempts":{"value":892},"def_orb":{"value":0.24660397074190177},"total_def_2pmid_made":{"value":390},"total_def_2prim_made":{"value":420},"off_2p":{"value":0.48764044943820223},"off_efg":{"value":0.5028268551236749},"off_to":{"value":0.19601930036188178},"total_off_fta":{"value":884},"def_efg":{"value":0.46285140562248994},"def_2prim":{"value":0.5675675675675675},"off_2pmid":{"value":0.40765765765765766},"def_2pmid":{"value":0.37142857142857144},"off_ftr":{"value":0.31236749116607776},"def_ftr":{"value":0.2817938420348059},"off_2prim":{"value":0.5672645739910314}},"off":{"doc_count":316,"total_off_fga":{"value":596},"total_def_2prim_attempts":{"value":178},"total_off_2pmid_attempts":{"value":198},"total_off_2pmid_made":{"value":68},"total_def_fta":{"value":212},"off_2pmidr":{"value":0.33221476510067116},"off_ppp":{"value":97.72727272727273},"off_3pr":{"value":0.3724832214765101},"total_def_fga":{"value":624},"def_poss":{"value":718},"total_off_2prim_made":{"value":88},"def_to":{"value":0.14763231197771587},"def_adj_opp":{"value":95.49005681818181},"off_3p":{"value":0.35135135135135137},"def_2p":{"value":0.44324324324324327},"off_2primr":{"value":0.2953020134228188},"def_2primr":{"value":0.28525641025641024},"points_scored":{"value":688},"off_orb":{"value":0.3177083333333333},"off_adj_opp":{"value":112.45097493036212},"total_def_2pmid_attempts":{"value":192},"def_ppp":{"value":105.84958217270194},"def_2pmidr":{"value":0.3076923076923077},"def_3pr":{"value":0.40705128205128205},"points_allowed":{"value":760},"def_3p":{"value":0.3858267716535433},"off_poss":{"value":704},"total_off_2prim_attempts":{"value":176},"def_orb":{"value":0.2864321608040201},"total_def_2pmid_made":{"value":60},"total_def_2prim_made":{"value":104},"off_2p":{"value":0.41711229946524064},"off_efg":{"value":0.45805369127516776},"off_to":{"value":0.21306818181818182},"total_off_fta":{"value":178},"def_efg":{"value":0.4983974358974359},"def_2prim":{"value":0.5842696629213483},"off_2pmid":{"value":0.3434343434343434},"def_2pmid":{"value":0.3125},"off_ftr":{"value":0.2986577181208054},"def_ftr":{"value":0.33974358974358976},"off_2prim":{"value":0.5}},"on":{"doc_count":746,"total_off_fga":{"value":2234},"total_def_2prim_attempts":{"value":562},"total_off_2pmid_attempts":{"value":690},"total_off_2pmid_made":{"value":294},"total_def_fta":{"value":630},"off_2pmidr":{"value":0.3088630259623993},"off_ppp":{"value":107.73353751914242},"off_3pr":{"value":0.37063563115487913},"total_def_fga":{"value":2364},"def_poss":{"value":2590},"total_off_2prim_made":{"value":418},"def_to":{"value":0.11814671814671815},"def_adj_opp":{"value":95.69494640122512},"off_3p":{"value":0.3526570048309179},"def_2p":{"value":0.45492957746478874},"off_2primr":{"value":0.3205013428827216},"def_2primr":{"value":0.2377326565143824},"points_scored":{"value":2814},"off_orb":{"value":0.3267776096822995},"off_adj_opp":{"value":113.16138996138996},"total_def_2pmid_attempts":{"value":858},"def_ppp":{"value":100.84942084942085},"def_2pmidr":{"value":0.3629441624365482},"def_3pr":{"value":0.3993231810490694},"points_allowed":{"value":2612},"def_3p":{"value":0.3008474576271186},"off_poss":{"value":2612},"total_off_2prim_attempts":{"value":716},"def_orb":{"value":0.23614775725593667},"total_def_2pmid_made":{"value":330},"total_def_2prim_made":{"value":316},"off_2p":{"value":0.5064011379800853},"off_efg":{"value":0.5147717099373321},"off_to":{"value":0.19142419601837674},"total_off_fta":{"value":706},"def_efg":{"value":0.45346869712351945},"def_2prim":{"value":0.5622775800711743},"off_2pmid":{"value":0.4260869565217391},"def_2pmid":{"value":0.38461538461538464},"off_ftr":{"value":0.3160250671441361},"def_ftr":{"value":0.26649746192893403},"off_2prim":{"value":0.5837988826815642}}}}},"status":200},{"took":18,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1068,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1062,"global_poss_count":{"value":3316},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":912,"poss_count":{"value":2956}},{"key":"Fernando, Bruno","doc_count":746,"poss_count":{"value":2612}},{"key":"Ayala, Eric","doc_count":732,"poss_count":{"value":2438}},{"key":"Smith, Jalen","doc_count":660,"poss_count":{"value":2308}},{"key":"Morsell, Darryl","doc_count":646,"poss_count":{"value":2290}},{"key":"Wiggins, Aaron","doc_count":696,"poss_count":{"value":1834}},{"key":"Smith Jr., Serrel","doc_count":380,"poss_count":{"value":934}},{"key":"Lindo Jr., Ricky","doc_count":412,"poss_count":{"value":916}},{"key":"Bender, Ivan","doc_count":72,"poss_count":{"value":176}},{"key":"Tomaic, Joshua","doc_count":46,"poss_count":{"value":104}},{"key":"Mona, Reese","doc_count":6,"poss_count":{"value":12}},{"key":"Ramsey, Trace","doc_count":2,"poss_count":{"value":0}}]}},"off":{"doc_count":316,"global_poss_count":{"value":704},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":264,"poss_count":{"value":596}},{"key":"Wiggins, Aaron","doc_count":236,"poss_count":{"value":542}},{"key":"Smith, Jalen","doc_count":226,"poss_count":{"value":492}},{"key":"Lindo Jr., Ricky","doc_count":218,"poss_count":{"value":460}},{"key":"Ayala, Eric","doc_count":180,"poss_count":{"value":394}},{"key":"Morsell, Darryl","doc_count":176,"poss_count":{"value":378}},{"key":"Smith Jr., Serrel","doc_count":160,"poss_count":{"value":374}},{"key":"Bender, Ivan","doc_count":68,"poss_count":{"value":174}},{"key":"Tomaic, Joshua","doc_count":44,"poss_count":{"value":98}},{"key":"Mona, Reese","doc_count":6,"poss_count":{"value":12}},{"key":"Ramsey, Trace","doc_count":2,"poss_count":{"value":0}}]}},"on":{"doc_count":746,"global_poss_count":{"value":2612},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Fernando, Bruno","doc_count":746,"poss_count":{"value":2612}},{"key":"Cowan, Anthony","doc_count":648,"poss_count":{"value":2360}},{"key":"Ayala, Eric","doc_count":552,"poss_count":{"value":2044}},{"key":"Morsell, Darryl","doc_count":470,"poss_count":{"value":1912}},{"key":"Smith, Jalen","doc_count":434,"poss_count":{"value":1816}},{"key":"Wiggins, Aaron","doc_count":460,"poss_count":{"value":1292}},{"key":"Smith Jr., Serrel","doc_count":220,"poss_count":{"value":560}},{"key":"Lindo Jr., Ricky","doc_count":194,"poss_count":{"value":456}},{"key":"Tomaic, Joshua","doc_count":2,"poss_count":{"value":6}},{"key":"Bender, Ivan","doc_count":4,"poss_count":{"value":2}}]}}}}},"status":200}]}
   ,
-  'autoOffQuery=false&baseQuery=NOT%20%28Terrell%20OR%20Valmon%20OR%20Clark%29&gender=Men&maxRank=400&minRank=0&offQuery=NOT%20Serrel%20AND%20Lindo%20AND%20Jalen&onQuery=NOT%20Serrel%20AND%20Lindo%20AND%20Bruno&team=Maryland&year=2018%2F9'
+  'game-autoOffQuery=false&baseQuery=NOT%20%28Terrell%20OR%20Valmon%20OR%20Clark%29&gender=Men&maxRank=400&minRank=0&offQuery=NOT%20Serrel%20AND%20Lindo%20AND%20Jalen&onQuery=NOT%20Serrel%20AND%20Lindo%20AND%20Bruno&team=Maryland&year=2018%2F9'
   :
   {"took":73,"responses":[{"took":73,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1424,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1400,"total_off_fga":{"value":3768},"total_def_2prim_attempts":{"value":1020},"total_off_2pmid_attempts":{"value":1136},"total_off_2pmid_made":{"value":466},"total_def_fta":{"value":1026},"off_2pmidr":{"value":0.30148619957537154},"off_ppp":{"value":108.55978260869566},"off_3pr":{"value":0.36889596602972397},"total_def_fga":{"value":3936},"def_poss":{"value":4406},"total_off_2prim_made":{"value":746},"def_to":{"value":0.1425329096686337},"def_adj_opp":{"value":98.6243206521739},"off_3p":{"value":0.3496402877697842},"def_2p":{"value":0.4459346186085499},"off_2primr":{"value":0.32961783439490444},"def_2primr":{"value":0.25914634146341464},"points_scored":{"value":4794},"off_orb":{"value":0.33753375337533753},"off_adj_opp":{"value":110.26268724466637},"total_def_2pmid_attempts":{"value":1366},"def_ppp":{"value":98.59282796187017},"def_2pmidr":{"value":0.3470528455284553},"def_3pr":{"value":0.3938008130081301},"points_allowed":{"value":4344},"def_3p":{"value":0.3174193548387097},"off_poss":{"value":4416},"total_off_2prim_attempts":{"value":1242},"def_orb":{"value":0.24425059476605868},"total_def_2pmid_made":{"value":504},"total_def_2prim_made":{"value":560},"off_2p":{"value":0.5096719932716569},"off_efg":{"value":0.5151273885350318},"off_to":{"value":0.19384057971014493},"total_off_fta":{"value":1222},"def_efg":{"value":0.4578252032520325},"def_2prim":{"value":0.5490196078431373},"off_2pmid":{"value":0.4102112676056338},"def_2pmid":{"value":0.3689604685212299},"off_ftr":{"value":0.3243099787685775},"def_ftr":{"value":0.2606707317073171},"off_2prim":{"value":0.6006441223832528}},"off":{"doc_count":106,"total_off_fga":{"value":200},"total_def_2prim_attempts":{"value":66},"total_off_2pmid_attempts":{"value":72},"total_off_2pmid_made":{"value":24},"total_def_fta":{"value":42},"off_2pmidr":{"value":0.36},"off_ppp":{"value":109.32203389830508},"off_3pr":{"value":0.32},"total_def_fga":{"value":232},"def_poss":{"value":250},"total_off_2prim_made":{"value":28},"def_to":{"value":0.128},"def_adj_opp":{"value":96.49491525423728},"off_3p":{"value":0.53125},"def_2p":{"value":0.4375},"off_2primr":{"value":0.32},"def_2primr":{"value":0.28448275862068967},"points_scored":{"value":258},"off_orb":{"value":0.3709677419354839},"off_adj_opp":{"value":110.0968},"total_def_2pmid_attempts":{"value":62},"def_ppp":{"value":104.8},"def_2pmidr":{"value":0.2672413793103448},"def_3pr":{"value":0.4482758620689655},"points_allowed":{"value":262},"def_3p":{"value":0.38461538461538464},"off_poss":{"value":236},"total_off_2prim_attempts":{"value":64},"def_orb":{"value":0.2328767123287671},"total_def_2pmid_made":{"value":22},"total_def_2prim_made":{"value":34},"off_2p":{"value":0.38235294117647056},"off_efg":{"value":0.515},"off_to":{"value":0.22033898305084745},"total_off_fta":{"value":68},"def_efg":{"value":0.5},"def_2prim":{"value":0.5151515151515151},"off_2pmid":{"value":0.3333333333333333},"def_2pmid":{"value":0.3548387096774194},"off_ftr":{"value":0.34},"def_ftr":{"value":0.1810344827586207},"off_2prim":{"value":0.4375}},"on":{"doc_count":134,"total_off_fga":{"value":276},"total_def_2prim_attempts":{"value":92},"total_off_2pmid_attempts":{"value":70},"total_off_2pmid_made":{"value":22},"total_def_fta":{"value":50},"off_2pmidr":{"value":0.2536231884057971},"off_ppp":{"value":98.17073170731707},"off_3pr":{"value":0.34057971014492755},"total_def_fga":{"value":308},"def_poss":{"value":348},"total_off_2prim_made":{"value":60},"def_to":{"value":0.15517241379310345},"def_adj_opp":{"value":97.37317073170732},"off_3p":{"value":0.3829787234042553},"def_2p":{"value":0.4854368932038835},"off_2primr":{"value":0.4057971014492754},"def_2primr":{"value":0.2987012987012987},"points_scored":{"value":322},"off_orb":{"value":0.2826086956521739},"off_adj_opp":{"value":109.53045977011494},"total_def_2pmid_attempts":{"value":114},"def_ppp":{"value":94.82758620689656},"def_2pmidr":{"value":0.37012987012987014},"def_3pr":{"value":0.33116883116883117},"points_allowed":{"value":330},"def_3p":{"value":0.3333333333333333},"off_poss":{"value":328},"total_off_2prim_attempts":{"value":112},"def_orb":{"value":0.20833333333333334},"total_def_2pmid_made":{"value":50},"total_def_2prim_made":{"value":50},"off_2p":{"value":0.45054945054945056},"off_efg":{"value":0.4927536231884058},"off_to":{"value":0.2073170731707317},"total_off_fta":{"value":78},"def_efg":{"value":0.4902597402597403},"def_2prim":{"value":0.5434782608695652},"off_2pmid":{"value":0.3142857142857143},"def_2pmid":{"value":0.43859649122807015},"off_ftr":{"value":0.2826086956521739},"def_ftr":{"value":0.16233766233766234},"off_2prim":{"value":0.5357142857142857}}}}},"status":200},{"took":23,"timed_out":false,"_shards":{"total":1,"successful":1,"skipped":0,"failed":0},"hits":{"total":{"value":1424,"relation":"eq"},"max_score":null,"hits":[]},"aggregations":{"tri_filter":{"buckets":{"baseline":{"doc_count":1400,"global_poss_count":{"value":4416},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Cowan, Anthony","doc_count":1174,"poss_count":{"value":3888}},{"key":"Fernando, Bruno","doc_count":932,"poss_count":{"value":3308}},{"key":"Ayala, Eric","doc_count":974,"poss_count":{"value":3204}},{"key":"Smith, Jalen","doc_count":820,"poss_count":{"value":2894}},{"key":"Morsell, Darryl","doc_count":812,"poss_count":{"value":2878}},{"key":"Wiggins, Aaron","doc_count":940,"poss_count":{"value":2644}},{"key":"Smith Jr., Serrel","doc_count":540,"poss_count":{"value":1360}},{"key":"Lindo Jr., Ricky","doc_count":558,"poss_count":{"value":1324}},{"key":"Bender, Ivan","doc_count":130,"poss_count":{"value":304}},{"key":"Tomaic, Joshua","doc_count":78,"poss_count":{"value":194}},{"key":"Mona, Reese","doc_count":40,"poss_count":{"value":82}},{"key":"Ramsey, Trace","doc_count":2,"poss_count":{"value":0}}]}},"off":{"doc_count":106,"global_poss_count":{"value":236},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Lindo Jr., Ricky","doc_count":106,"poss_count":{"value":236}},{"key":"Smith, Jalen","doc_count":106,"poss_count":{"value":236}},{"key":"Cowan, Anthony","doc_count":104,"poss_count":{"value":232}},{"key":"Ayala, Eric","doc_count":82,"poss_count":{"value":184}},{"key":"Morsell, Darryl","doc_count":64,"poss_count":{"value":142}},{"key":"Wiggins, Aaron","doc_count":64,"poss_count":{"value":142}},{"key":"Mona, Reese","doc_count":2,"poss_count":{"value":8}},{"key":"Fernando, Bruno","doc_count":2,"poss_count":{"value":0}}]}},"on":{"doc_count":134,"global_poss_count":{"value":328},"player":{"doc_count_error_upper_bound":0,"sum_other_doc_count":0,"buckets":[{"key":"Fernando, Bruno","doc_count":134,"poss_count":{"value":328}},{"key":"Lindo Jr., Ricky","doc_count":134,"poss_count":{"value":328}},{"key":"Cowan, Anthony","doc_count":120,"poss_count":{"value":304}},{"key":"Ayala, Eric","doc_count":110,"poss_count":{"value":256}},{"key":"Morsell, Darryl","doc_count":90,"poss_count":{"value":220}},{"key":"Wiggins, Aaron","doc_count":80,"poss_count":{"value":204}},{"key":"Smith, Jalen","doc_count":2,"poss_count":{"value":0}}]}}}}},"status":200}]}
 };
